Memoize TimelineEvent
