+++ oarepo_requests/ui/theme/assets/semantic-ui/js/oarepo_requests_ui/request-detail/components/index.js
@@ -5,4 +5,5 @@ export { ConfirmModal } from './ConfirmModal';
 export { MainRequestDetails } from './MainRequestDetails';
 export { TopicPreview } from './TopicPreview';
 export { Timeline } from './Timeline';
+export { default as TimelineEvent } from './TimelineEvent';
 export { EventSubmitForm } from './EventSubmitForm';
